Finished A/B test for SP html vs pdf
