+++ packages/numenta.com/pages/resources/biological-and-machine-intelligence/spatial-pooling-algorithm/index.css
